requirements.txt
asgiref==3.7.2
boto3==1.26.161
botocore==1.29.161
cssselect2==0.7.0
Django==4.2.1
django-classy-tags==4.0.0
django-cms==3.11.3
django-dotenv==1.4.2
django-filer==3.0.0
django-formtools==2.4.1
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-storages==1.13.2
django-treebeard==4.4
djangocms-admin-style==3.2.4
djangocms-attributes-field==3.0.0
djangocms-file==3.0.1
djangocms-googlemap==2.1.0
djangocms-link==3.1.1
djangocms-picture==4.0.0
djangocms-snippet==3.1.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.1.3
djangocms-video==3.0.0
easy-thumbnails==2.8.5
html5lib==1.1
jmespath==1.0.1
lxml==4.9.3
nanoid==2.0.0
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
reportlab==4.0.4
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
webencodings==0.5.1
